box spawn 완성 + 강화로직 체인지 완료

+++ Assets/Safety Hero/Scripts/Manager/GlobalManager.cs
@@ -21,7 +21,7 @@ public class GlobalManager : MonoBehaviour
 {
     public static GlobalManager instance;
     private PlayerInput playerInput;
-    public int nowTimeScale = 1;
+    public float nowTimeScale = 1;
     public PlayerState playerState;
     public GameState gameState;
